+++ examples/1INFILTR/validation.py
@@ -1,6 +1,11 @@
 #!/usr/bin/env python
 
-"""1INFILTR case from Hydrus-1D, horizontal."""
+"""
+1INFILTR case from Hydrus-1D in horizontal orientation.
+
+Data obtained using Hydrus-1D version 4.17.0140
+1001 nodes, water content tolerance = 1e-6
+"""
 
 from __future__ import division, absolute_import, print_function